reorganize/rename PPMS facilty directories & files
