+++ config/initializers/breakers.rb
@@ -50,7 +50,7 @@ services = [
   VIC::Configuration.instance.breakers_service,
   Facilities::AccessWaitTimeConfiguration.instance.breakers_service,
   Facilities::AccessSatisfactionConfiguration.instance.breakers_service,
-  Facilities::PPMSConfiguration.instance.breakers_service,
+  Facilities::PPMS::Configuration.instance.breakers_service,
   VIC::Configuration.instance.breakers_service,
   GI::Configuration.instance.breakers_service,
   HCA::Configuration.instance.breakers_service,
