+++ spec/lib/facilities/ppms_client_spec.rb
@@ -2,11 +2,11 @@
 
 require 'rails_helper'
 module Facilities
-  RSpec.describe PPMSClient do
+  RSpec.describe PPMS::Client do
     let(:bbox_bounds) { [-79, 38, -77, 39] }
 
-    it 'is an PPMSClient object' do
-      expect(described_class.new).to be_an(PPMSClient)
+    it 'is an PPMS::Client object' do
+      expect(described_class.new).to be_an(PPMS::Client)
     end
 
     regex_matcher = lambda { |r1, r2|
@@ -21,7 +21,7 @@ module Facilities
 
         it 'logs an error and raise GatewayTimeout' do
           expect do
-            PPMSClient.new.provider_locator('bbox': bbox_bounds)
+            PPMS::Client.new.provider_locator('bbox': bbox_bounds)
           end.to raise_error(Common::Exceptions::GatewayTimeout)
         end
       end
@@ -29,7 +29,7 @@ module Facilities
       context 'with an unknown error from PPMS' do
         it 'raises BackendUnhandledException when errors happen' do
           VCR.use_cassette('facilities/va/ppms_500', match_requests_on: [regex_matcher]) do
-            expect { PPMSClient.new.provider_locator('bbox': bbox_bounds) }
+            expect { PPMS::Client.new.provider_locator('bbox': bbox_bounds) }
               .to raise_error(Common::Exceptions::BackendServiceException) do |e|
                 expect(e.message).to match(/PPMS_502/)
               end
@@ -39,7 +39,7 @@ module Facilities
 
       it 'finds at least one provider' do
         VCR.use_cassette('facilities/va/ppms', match_requests_on: [regex_matcher]) do
-          r = PPMSClient.new.provider_locator('bbox': [-79, 38, -77, 39])
+          r = PPMS::Client.new.provider_locator('bbox': [-79, 38, -77, 39])
           expect(r.length).to be > 0
           expect(r[0]['Latitude']).to be > 38
           expect(r[0]['Latitude']).to be < 39
@@ -48,7 +48,7 @@ module Facilities
 
       it 'returns a Provider shape' do
         VCR.use_cassette('facilities/va/ppms', match_requests_on: [regex_matcher]) do
-          r = PPMSClient.new.provider_info(1_427_435_759)
+          r = PPMS::Client.new.provider_info(1_427_435_759)
           expect(r['ProviderIdentifier']).not_to be(nil)
           expect(r['Name']).not_to be(nil)
           expect(r['ProviderSpecialties'].class.name).to eq('Array')
@@ -57,7 +57,7 @@ module Facilities
 
       it 'returns some Specialties' do
         VCR.use_cassette('facilities/va/ppms', match_requests_on: [regex_matcher]) do
-          r = PPMSClient.new.specialties
+          r = PPMS::Client.new.specialties
           expect(r.length).to be > 0
           expect(r[0]['SpecialtyCode']).not_to be(nil)
         end
@@ -65,7 +65,7 @@ module Facilities
 
       it 'returns a CareSite' do
         VCR.use_cassette('facilities/va/ppms', match_requests_on: [regex_matcher]) do
-          r = PPMSClient.new.provider_caresites(1_427_435_759)
+          r = PPMS::Client.new.provider_caresites(1_427_435_759)
           expect(r.length).to be > 0
           expect(r[0]['Longitude']).not_to be(nil)
         end
@@ -73,7 +73,7 @@ module Facilities
 
       it 'returns Services' do
         VCR.use_cassette('facilities/va/ppms', match_requests_on: [regex_matcher]) do
-          r = PPMSClient.new.provider_services(1_427_435_759)
+          r = PPMS::Client.new.provider_services(1_427_435_759)
           expect(r.length).to be > 0
           expect(r[0]['Longitude']).not_to be(nil)
           expect(r[0]['CareSiteAddressStreet']).not_to be(nil)
@@ -81,7 +81,7 @@ module Facilities
       end
 
       it 'edits the parameters' do
-        params = PPMSClient.new.build_params('bbox': [73, -60, 74, -61])
+        params = PPMS::Client.new.build_params('bbox': [73, -60, 74, -61])
         Rails.logger.info(params)
         expect(params[:radius]).to be > 35
       end
